+++ crates/matrix-sdk-crypto/src/verification/requests.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{sync::Arc, time::Duration};
+use std::{ops::Add, sync::Arc, time::Duration};
 
 use eyeball::shared::{Observable as SharedObservable, ObservableWriteGuard, WeakObservable};
 use futures_core::Stream;
@@ -321,6 +321,15 @@ impl VerificationRequest {
     /// Has the verification flow timed out.
     pub fn timed_out(&self) -> bool {
         self.creation_time.elapsed() > VERIFICATION_TIMEOUT
+    }
+
+    /// Get the time left before the verification flow will time out, without
+    /// further action.
+    pub fn time_remaining(&self) -> Duration {
+        self.creation_time
+            .add(VERIFICATION_TIMEOUT)
+            .checked_duration_since(Instant::now())
+            .unwrap_or(Duration::from_secs(0))
     }
 
     /// Get the supported verification methods of the other side.
@@ -1590,7 +1599,10 @@ struct Done {}
 #[cfg(test)]
 mod tests {
 
-    use std::convert::{TryFrom, TryInto};
+    use std::{
+        convert::{TryFrom, TryInto},
+        time::Duration,
+    };
 
     use assert_matches::assert_matches;
     #[cfg(feature = "qrcode")]
@@ -1639,6 +1651,8 @@ mod tests {
         );
 
         assert_matches!(bob_request.state(), VerificationRequestState::Created { .. });
+        assert!(bob_request.time_remaining() <= Duration::from_secs(600)); // 10 minutes
+        assert!(bob_request.time_remaining() > Duration::from_secs(540)); // 9 minutes
 
         #[allow(clippy::needless_borrow)]
         let alice_request = VerificationRequest::from_request(